+++ src/Bridge/IBridgeMessage.ts
@@ -1,0 +1,4 @@
+export default interface IBridgeMessage<TMessage> {
+	invocationUid: string;
+	message: TMessage;
+}