+++ test/integration/Bridge/bridgeManagement.ts
@@ -1,11 +1,14 @@
 import * as sinon from 'sinon';
 import BridgeServer from '../../../src/Bridge/BridgeServer';
 import BridgeClient from '../../../src/Bridge/BridgeClient';
+import { createMockSocketServerClientPair } from '../WebSocket/mockWebSocket';
 
 const params = require('../../../config/parameters');
 
-export function createBridgeAndItsDependencies() {
-	const fileSystem = {};
+export async function createBridgeAndItsDependencies() {
+	const fileSystem = {
+		onStorageUnitsChanged: sinon.spy(),
+	};
 	const fileDetailsProvider = {};
 	const nativeDriver = {};
 	const display = {};
@@ -19,6 +22,13 @@ export function createBridgeAndItsDependencies() {
 	const cecListener = {
 		listen: sinon.stub().resolves(),
 		close: sinon.stub().resolves(),
+		onKeypress: sinon.spy(),
+	};
+	const { socketServer, socketClient } = createMockSocketServerClientPair();
+	const socketServerWrapper = {
+		server: socketServer,
+		listen: () => Promise.resolve(),
+		close: () => Promise.resolve(),
 	};
 	const bridgeServer = new BridgeServer(
 		params.server.bridge_url,
@@ -29,8 +39,9 @@ export function createBridgeAndItsDependencies() {
 		videoPlayer as any,
 		overlayRenderer as any,
 		cecListener as any,
+		() => (socketServerWrapper as any),
 	);
-	const bridgeClient = new BridgeClient(params.server.bridge_url);
+	const bridgeClient = new BridgeClient(params.server.bridge_url, socketClient);
 	return {
 		fileSystem,
 		fileDetailsProvider,