+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Support PIP (only fullscreen)
 - Support proximity sensor on NEC displays (there it's called "human sensor")
 - Support reset NEC display settings as means of recovery from breaking external changes
+- Internal client-server communication re-implemented using websocket instead of HTTP requests
 
 ### Fixed
 - Casually not loading of applet after restart app